Complete project 5 challenges
Challenge
One of the best ways to learn is to write your own code as often as
possible, so here are three ways you should try extending this app to
make sure you fully understand what’s going on:

1. Disallow answers that are shorter than three letters or are just
our start word. For the three- letter check, the easiest thing to do
is put a check into isReal() that returns false if the word length is
under three letters. For the second part, just compare the start word
against their input word and return false if they are the same.

2. Refactor all the else statements we just added so that they call a
new method called showErrorMessage(). This should accept an error
message and a title, and do all the UIAlertController work from there.

3. Add a left bar button item that calls startGame(), so users can
restart with a new word whenever they want to.